pyproject.toml
[tool.poetry]
name = "authorization-service"
version = "0.1.0"
description = ""
authors = ["NeZlox <cnegovek@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1"
python-crontab = "^3"
litestar = { extras = ["standard"], version = "^2" }
litestar-granian = "^0"
litestar-asyncpg = "^0"
structlog = "^24"
advanced-alchemy = "^0.20"
sqlalchemy-utils = "^0"

aiohttp = "^3.11"

ruff = "^0.11"

# personal_growth_sdk = { git = "https://github.com/NeZlox/personal-growth-sdk.git", branch = "main" }
personal_growth_sdk = { path = "../personal-growth-sdk", develop = true }
passlib = { version = "^1.7.4", extras = ["argon2"] }
python-jose = "^3.4.0"
multidict = "6.2.0"
cryptography = "^44.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
